requirements.txt
aiodns==3.5.0
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-socks==0.9.4
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.4.0
backcall==0.2.0
bandit==1.8.6
beautifulsoup4==4.14.2
black==25.9.0
bleach==6.2.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
h11==0.16.0
holidays==0.82
identify==2.6.15
idna==3.11
iniconfig==2.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
magic-filter==1.0.12
mando==0.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nodeenv==1.9.1
numpy==2.3.4
ollama==0.6.1
outcome==1.3.0.post0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pbr==7.0.1
pickleshare==0.7.5
pillow==12.0.0
pip-review==1.3.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
pure_eval==0.2.3
pycares==4.11.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
pyupgrade==3.21.0
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
radon==6.0.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
ruff==0.14.1
scheduler==0.8.8
selenium==4.36.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
stevedore==5.5.0
telegram==0.0.1
tinycss2==1.4.0
tokenize_rt==6.2.0
tornado==6.5.2
tox-ini-fmt==1.7.0
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
typeguard==4.4.4
types-pytz==2025.2.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
virtualenv==20.35.3
vulture==2.14
wcwidth==0.2.14
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
wsproto==1.2.0
yarg==0.1.10
yarl==1.22.0
